fix: Fix calculation of AABB for `ShapeHitbox`es (#1481)
