+++ packages/flame/lib/src/collisions/hitboxes/shape_hitbox.dart
@@ -40,6 +40,7 @@ mixin ShapeHitbox on ShapeComponent implements Hitbox<ShapeHitbox> {
   late Function() _transformListener;
 
   final Vector2 _halfExtents = Vector2.zero();
+  static const double _extentEpsilon = 0.000000000000001;
   final Matrix3 _rotationMatrix = Matrix3.zero();
 
   @override
@@ -134,9 +135,12 @@ mixin ShapeHitbox on ShapeComponent implements Hitbox<ShapeHitbox> {
 
   Aabb2 _recalculateAabb() {
     final size = absoluteScaledSize;
-    // This has +1 since a point on the edge of the bounding box is currently
-    // counted as outside.
-    _halfExtents.setValues(size.x + 1, size.y + 1);
+    // This has double.minPositive since a point on the edge of the AABB is
+    // currently counted as outside.
+    _halfExtents.setValues(
+      size.x / 2 + _extentEpsilon,
+      size.y / 2 + _extentEpsilon,
+    );
     _rotationMatrix.setRotationZ(absoluteAngle);
     _validAabb = true;
     return _aabb
